refactor(pretty): break into submodules
