+++ src/tools/pretty/spinners.ts
@@ -1,0 +1,37 @@
+import { print } from "gluegun"
+import { INDENT } from "."
+
+const { gray } = print.colors
+
+type Spinner = ReturnType<typeof print.spin>
+const spinners: { [key: string]: Spinner } = {}
+
+export const startSpinner = (m = "") => {
+  let spinner = spinners[m]
+  if (!spinner) {
+    spinner = print.spin({ prefixText: INDENT, text: gray(m) })
+    spinners[m] = spinner
+  }
+  return spinner
+}
+
+export const stopSpinner = (m: string, symbol: string) => {
+  const spinner = spinners[m]
+  if (spinner) {
+    spinner.stopAndPersist({ symbol })
+    delete spinners[m]
+  }
+}
+
+export const clearSpinners = () => {
+  Object.keys(spinners).forEach((m) => {
+    spinners[m].stop()
+    delete spinners[m]
+  })
+}
+
+export const spinner = {
+  start: startSpinner,
+  stop: stopSpinner,
+  clear: clearSpinners,
+} as const
